[CL 14233282 by Marc Audy in ue5-main branch]

+++ Engine/Source/Programs/AutomationTool/Gauntlet/Unreal/Base/Gauntlet.UnrealPGONode.cs
@@ -6,6 +6,8 @@ using System.Collections.Generic;
 using AutomationTool;
 using UnrealBuildTool;
 using System.Diagnostics;
+using System.Diagnostics.Eventing.Reader;
+using Gauntlet.Utils;
 
 namespace Gauntlet
 {
@@ -143,32 +145,45 @@ namespace Gauntlet
 			{
 				ScreenshotTime = DateTime.Now;
 
-				try
+				if (!File.Exists(Path.Combine(ScreenshotDirectory, ImageFilename)))
 				{
-					TimeSpan ImageTimestamp = DateTime.UtcNow - ScreenshotStartTime;
-					string ImageOutputPath = Path.Combine(ScreenshotDirectory, ImageTimestamp.ToString().Replace(':', '-') + ".jpg");
-					ImageUtils.ResaveImageAsJpgWithScaleAndQuality(Path.Combine(ScreenshotDirectory, ImageFilename), ImageOutputPath, ScreenshotScale, ScreenshotQuality);
+					Log.Info("PGOPlatform.TakeScreenshot returned true, but output image {0} does not exist! skipping", ImageFilename);
+				}
+				else if(new FileInfo(Path.Combine(ScreenshotDirectory, ImageFilename)).Length <= 0)
+				{
+					Log.Info("PGOPlatform.TakeScreenshot returned true, but output image {0} is size 0! skipping", ImageFilename);
+				}
+				else
+				{
+					try
+					{
+						TimeSpan ImageTimestamp = DateTime.UtcNow - ScreenshotStartTime;
+						string ImageOutputPath = Path.Combine(ScreenshotDirectory, ImageTimestamp.ToString().Replace(':', '-') + ".jpg");
+						ImageUtils.ResaveImageAsJpgWithScaleAndQuality(Path.Combine(ScreenshotDirectory, ImageFilename), ImageOutputPath, ScreenshotScale, ScreenshotQuality);
 
-					// Delete the temporary image file
-					try { File.Delete(Path.Combine(ScreenshotDirectory, ImageFilename)); }
+						// Delete the temporary image file
+						try { File.Delete(Path.Combine(ScreenshotDirectory, ImageFilename)); }
+						catch (Exception e)
+						{
+							Log.Warning("Got Exception Deleting temp iamge: {0}", e.ToString());
+						}
+					}
 					catch (Exception e)
 					{
-						Log.Warning("Got Exception Deleting temp iamge: {0}", e.ToString());
+						Log.Info("Got Exception Renaming PGO image {0}: {1}", ImageFilename, e.ToString());
+
+						TimeSpan ImageTimestamp = DateTime.UtcNow - ScreenshotStartTime;
+						string CopyFileName = Path.Combine(ScreenshotDirectory, ImageTimestamp.ToString().Replace(':', '-') + ".bmp");
+						Log.Info("Copying unconverted image {0} to {1}", ImageFilename, CopyFileName);
+						try
+						{
+							File.Copy(Path.Combine(ScreenshotDirectory, ImageFilename), CopyFileName);
+						}
+						catch (Exception e2)
+						{
+							Log.Warning("Got Exception copying un-converted screenshot image: {0}", e2.ToString());
+						}
 					}
-				}
-				catch(Exception e)
-				{
-					Log.Warning("Got Exception Renaming PGO image {0}: {1}", ImageFilename, e.ToString());
-					Process proc = Process.GetCurrentProcess();
-					Log.Info("Memory Usage: Private: {0}", proc.PrivateMemorySize64);
-					Log.Info("Memory Usage: Virtual: {0}", proc.VirtualMemorySize64);
-					Log.Info("Memory Usage: Peak Virtual: {0}", proc.PeakVirtualMemorySize64);
-					Log.Info("Memory Usage: Paged: {0}", proc.PagedMemorySize64);
-					Log.Info("Memory Usage: System Paged: {0}", proc.PagedSystemMemorySize64);
-					Log.Info("Memory Usage: System NonPaged: {0}", proc.NonpagedSystemMemorySize64);
-					Log.Info("Memory Usage: Working Set: {0}", proc.WorkingSet64);
-					Log.Info("Memory Usage: Peak Working Set: {0}", proc.PeakWorkingSet64);
-					proc.Dispose();
 				}
 			}
